Setup the bridge networking.
TODO|NOTE:
This can be "dangerous". If we don't get "br-physical" up and running,
everything else here will fail!!  BUT (!) it seems that MAYBE (!) it
works fine IF (!) doing from scratch.
One of these days I need to actually test it, but I'm afraid to "rock
the boat" now. Everything almost works! :)